Add EventType to API
