+++ src/aiohubspace/types.py
@@ -1,0 +1,12 @@
+from enum import Enum
+
+
+class EventType(Enum):
+    """Enum with possible Events."""
+
+    RESOURCE_ADDED = "add"
+    RESOURCE_UPDATED = "update"
+    RESOURCE_DELETED = "delete"
+    CONNECTED = "connected"
+    DISCONNECTED = "disconnected"
+    RECONNECTED = "reconnected"
